requirement.txt
add-trailing-comma==2.4.0
aenum==3.1.11
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autoflake==1.5.3
black==22.12.0
camera==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
ConfigArgParse==1.5.3
ecdsa==0.18.0
fixit==0.1.4
flake8==6.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
flask-open-directory==0.1.1
flatdict==4.0.1
frozenlist==1.3.3
gray==0.13.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
ldap3==2.9.1
libcst==0.4.9
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.1
okta==2.8.0
opencv-python==4.6.0.66
pathspec==0.10.3
pkgutil_resolve_name==1.3.10
platformdirs==2.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycryptodome==3.17
pydash==6.0.0
pyflakes==3.0.1
Pygments==2.13.0
pyrsistent==0.19.2
python-jose==3.3.0
pyupgrade==3.3.1
PyYAML==6.0
pyzbar==0.1.9
rich==12.6.0
rsa==4.9
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
six==1.16.0
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
unify==0.5
untokenize==0.1.1
Werkzeug==2.2.2
xmltodict==0.13.0
yaml-1.3==0.1.0
yarl==1.8.2
zipp==3.11.0
add-trailing-comma==2.4.0
aenum==3.1.11
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autoflake==1.5.3
black==22.12.0
camera==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
ConfigArgParse==1.5.3
ecdsa==0.18.0
fixit==0.1.4
flake8==6.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
flask-open-directory==0.1.1
flatdict==4.0.1
frozenlist==1.3.3
gray==0.13.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
ldap3==2.9.1
libcst==0.4.9
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.1
okta==2.8.0
opencv-python==4.6.0.66
pathspec==0.10.3
pkgutil_resolve_name==1.3.10
platformdirs==2.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycryptodome==3.17
pydash==6.0.0
pyflakes==3.0.1
Pygments==2.13.0
pyrsistent==0.19.2
python-jose==3.3.0
pyupgrade==3.3.1
PyYAML==6.0
pyzbar==0.1.9
rich==12.6.0
rsa==4.9
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
six==1.16.0
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
unify==0.5
untokenize==0.1.1
Werkzeug==2.2.2
xmltodict==0.13.0
yaml-1.3==0.1.0
yarl==1.8.2
zipp==3.11.0
